refactor: decompose 4 more monolith files into submodules (issue #2486 sprint 5)
- c3d_reader.py (917 → 480 LOC) → _c3d_io, _c3d_models, _c3d_force_plates, _c3d_export
- swing_optimizer.py (918 → 504 LOC) → _swing_models, _swing_objectives,
  _swing_constraints, _swing_kinematics
- service.py (915 → 483 LOC) → _gas_flow, _concentration, _heating_value,
  _performance, _errors
- chain_manipulation.py (952 → 18 LOC) → _chain_model, _chain_widget,
  _chain_visualizer, _chain_dialogs

All public symbols remain importable from original module paths.
